Moved demo things into app folder. Refactored data download script. Still lots of work to do.
