+++ use_cases/eluc/.gitignore
@@ -7,11 +7,13 @@ experiments/predictor_significance
 # Ignores figures for paper
 experiments/figures
 
-experiments/figures
-
 # Ignores trained prescriptors and seeds
 prescriptors/*/trained_prescriptors
 prescriptors/*/seeds
 
 data/*.zip
+data/processed/*.csv
 *.nc
+
+# Ignore demo folder
+demo/
