+++ use_cases/eluc/app/assets/favicon.ico
