--- use_cases/eluc/demo/predictors/predictors.json
@@ -1,22 +1,0 @@
-{"predictors": [
-    {
-        "name": "Linear Regression (West Europe)",
-        "filename": "143_linear.joblib"
-    },
-    {
-        "name": "Linear Regression (USA)",
-        "filename": "4_linear.joblib"
-    },
-    {
-        "name": "Linear Regression (Brazil)",
-        "filename": "29_linear.joblib"
-    },
-    {
-        "name": "Linear Regression (General)",
-        "filename": "ELUC_linear.joblib"
-    },
-    {
-        "name": "Random Forest (West Europe)",
-        "filename": "ELUC_forest.joblib"
-    }
-]}
