--- use_cases/eluc/demo/prescriptors/pareto.csv
@@ -1,14 +1,0 @@
-id,identity,ELUC,NSGA-II_crowding_distance,NSGA-II_rank,change,is_elite
-92_70,"{'ancestor_count': 6, 'ancestor_ids': ['90_54', '84_43'], 'birth_generation': 92, 'domain_name': None, 'experiment_version': 'LinearLeafLandUseDecode', 'unique_id': '92_70', 'origin': '90_54~CUW~84_43#MGNP'}",0.0182805009466933,inf,1,0.0667467755018339,True
-99_65,"{'ancestor_count': 11, 'ancestor_ids': ['98_81', '98_81'], 'birth_generation': 99, 'domain_name': None, 'experiment_version': 'LinearLeafLandUseDecode', 'unique_id': '99_65', 'origin': '98_81~CUW~98_81#MGNP'}",-5.987312544991907,0.3791044718989205,1,0.1012967088720754,True
-97_97,"{'ancestor_count': 95, 'ancestor_ids': ['96_14', '96_88'], 'birth_generation': 97, 'domain_name': None, 'experiment_version': 'LinearLeafLandUseDecode', 'unique_id': '97_97', 'origin': '96_14~CUW~96_88#MGNP'}",-15.413083904594863,0.3229220023442851,1,0.1719288420747773,True
-99_51,"{'ancestor_count': 98, 'ancestor_ids': ['98_49', '98_66'], 'birth_generation': 99, 'domain_name': None, 'experiment_version': 'LinearLeafLandUseDecode', 'unique_id': '99_51', 'origin': '98_49~CUW~98_66#MGNP'}",-24.47153974309113,0.1251737620298189,1,0.2249340374604773,False
-100_91,"{'ancestor_count': 99, 'ancestor_ids': ['99_15', '99_13'], 'birth_generation': 100, 'domain_name': None, 'experiment_version': 'LinearLeafLandUseDecode', 'unique_id': '100_91', 'origin': '99_15~CUW~99_13#MGNP'}",-28.439043897759447,0.1606969193537109,1,0.245644698452457,True
-100_58,"{'ancestor_count': 99, 'ancestor_ids': ['99_58', '99_48'], 'birth_generation': 100, 'domain_name': None, 'experiment_version': 'LinearLeafLandUseDecode', 'unique_id': '100_58', 'origin': '99_58~CUW~99_48#MGNP'}",-33.73895940604969,0.1366667417253438,1,0.2842010471172498,False
-100_96,"{'ancestor_count': 99, 'ancestor_ids': ['99_38', '99_51'], 'birth_generation': 100, 'domain_name': None, 'experiment_version': 'LinearLeafLandUseDecode', 'unique_id': '100_96', 'origin': '99_38~CUW~99_51#MGNP'}",-35.40003462234897,0.154603585947419,1,0.2927465910239528,True
-100_40,"{'ancestor_count': 99, 'ancestor_ids': ['99_39', '99_58'], 'birth_generation': 100, 'domain_name': None, 'experiment_version': 'LinearLeafLandUseDecode', 'unique_id': '100_40', 'origin': '99_39~CUW~99_58#MGNP'}",-39.91197247797104,0.124807711912608,1,0.3302908233641747,False
-100_29,"{'ancestor_count': 99, 'ancestor_ids': ['99_13', '99_21'], 'birth_generation': 100, 'domain_name': None, 'experiment_version': 'LinearLeafLandUseDecode', 'unique_id': '100_29', 'origin': '99_13~CUW~99_21#MGNP'}",-44.0336422258556,0.1583452126379033,1,0.3635849260809409,True
-100_54,"{'ancestor_count': 99, 'ancestor_ids': ['99_78', '99_38'], 'birth_generation': 100, 'domain_name': None, 'experiment_version': 'LinearLeafLandUseDecode', 'unique_id': '100_54', 'origin': '99_78~CUW~99_38#MGNP'}",-59.63589819674576,0.1512997596627784,1,0.4870467596605408,True
-100_100,"{'ancestor_count': 99, 'ancestor_ids': ['99_85', '99_78'], 'birth_generation': 100, 'domain_name': None, 'experiment_version': 'LinearLeafLandUseDecode', 'unique_id': '100_100', 'origin': '99_85~CUW~99_78#MGNP'}",-65.90170416132273,0.149356051396792,1,0.5662291414651899,True
-99_78,"{'ancestor_count': 98, 'ancestor_ids': ['98_15', '98_78'], 'birth_generation': 99, 'domain_name': None, 'experiment_version': 'LinearLeafLandUseDecode', 'unique_id': '99_78', 'origin': '98_15~CUW~98_78#MGNP'}",-66.05238226411777,0.141290265883081,1,0.6159501550968247,True
-99_39,"{'ancestor_count': 98, 'ancestor_ids': ['98_66', '93_62'], 'birth_generation': 99, 'domain_name': None, 'experiment_version': 'LinearLeafLandUseDecode', 'unique_id': '99_39', 'origin': '98_66~CUW~93_62#MGNP'}",-68.41783872386198,inf,1,0.7514420561129526,True
